requirements.txt
flask
click
frozen-flask
mistune>=2a
python-slugify
pyyaml
pygments
pytest
feedgen

